refactor: modularize dashboard features
